% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data frame with three variables that contain same-sized vectors describing nodes, edges and wieghts of a graph}

\item{init_node}{The node from which the shortest distance to all other nodes is calculated}
}
\value{
A vector of the distances from init_node to all other nodes
}
\description{
Implementation of Djikstra's algorithm to find the shortest path to all nodes from an initial node using an adjacency matrix
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstras_algorithm} and \url{https://www.youtube.com/watch?v=Ww4nU_NcAIQ}
}
