\name{nicolasyuki-package}
\alias{nicolasyuki-package}
\alias{nicolasyuki}
\docType{package}
\title{
\packageTitle{nicolasyuki}
}
\description{
\packageDescription{nicolasyuki}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{nicolasyuki}
\packageIndices{nicolasyuki}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{nicolasyuki}

Maintainer: \packageMaintainer{nicolasyuki}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
