% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.r
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{wiki_graph data}
\format{
The \code{data.frame} contains 3 variables:
\describe{
  \item{v1}{nodes}
  \item{v2}{nodes}
  \item{w}{weights between the nodes}
}
}
\usage{
wiki_graph
}
\description{
wiki_graph: DataFrame containing three columns (v1, v2, w) and 18 entries.
}
\references{

}
\keyword{datasets}
