\name{nicolasyuki-package}
\alias{nicolasyuki-package}
\alias{nicolasyuki}
\docType{package}
\title{
\packageTitle{nicolasyuki}
}
\description{
\packageDescription{nicolasyuki}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{nicolasyuki}
\packageIndices{nicolasyuki}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{nicolasyuki}

Maintainer: \packageMaintainer{nicolasyuki}
}
